+++ NEWS.md
@@ -6,6 +6,7 @@
 * Fixed plotting issue in `warpRefMesh`
 * Fixed single variable issue in `compare.evol.rates` related to R 4.0.2 changes
 * Added factor labels to output object CR.mat in `modularity.test` & `phylo.modularity`
+* Fixed bug in computations of `phylo.modularity`
 * Fixed order consistency for pairwise calculations in `phylo.modularity` 
 
 # CHANGES IN GEOMORPH VERSION 3.3.1 (Patch Release)